+++ src/Encamina.Enmarcha.AspNet.OpenApi/Constants.cs
@@ -1,0 +1,18 @@
+﻿namespace Encamina.Enmarcha.AspNet.OpenApi;
+
+/// <summary>
+/// Constants used in the OpenAPI library.
+/// </summary>
+public static class Constants
+{
+    /// <summary>
+    /// OpenAPI related constants.
+    /// </summary>
+    public static class OpenApi
+    {
+        /// <summary>
+        /// The OpenAPI key header.
+        /// </summary>
+        public static readonly string KeyHeader = @"x-openapi-key";
+    }
+}